+++ courses.js
@@ -1,11 +1,21 @@
 // Put courses you want in this array
 
 module.exports = [
+  // Beginner path - Vue 2
+  // "https://www.vuemastery.com/courses/intro-to-vue-js/list-rendering",
+  // "https://www.vuemastery.com/courses/real-world-vue-js/real-world-intro",
+  // "https://www.vuemastery.com/courses/mastering-vuex/intro-to-vuex",
+  // "https://www.vuemastery.com/courses/next-level-vue/next-level-vue-orientation",
+  // "https://www.vuemastery.com/courses/unit-testing/what-to-test",
+
+  // Beginner path - Vue 3
   // "https://www.vuemastery.com/courses/intro-to-vue-3/creating-the-vue-app-vue3",
   // "https://www.vuemastery.com/courses/real-world-vue3/rwv3-orientation",
   // "https://www.vuemastery.com/courses/touring-vue-router/vue-router-introduction",
   // "https://www.vuemastery.com/courses/vuex-fundamentals/vuex4-intro-to-vuex",
   // "https://www.vuemastery.com/courses/unit-testing-vue-3/what-to-test-vue-3/",
+
+  // Intermediate path
   // "https://www.vuemastery.com/courses/real-world-testing/getting-started",
   // "https://www.vuemastery.com/courses/vue3-forms/forms-introduction",
   // "https://www.vuemastery.com/courses/validating-vue3-forms/why-vee-validate",
@@ -16,6 +26,8 @@ module.exports = [
   // "https://www.vuemastery.com/courses/scaling-vue-with-nuxt-js/why-use-nuxt",
   // "https://www.vuemastery.com/courses/beautify-with-vuetify/getting-started-with-vuetify",
   // "https://www.vuemastery.com/courses/progressive-web-apps-vue-3/what-are-progressive-web-apps",
+
+  // Advanced path
   // "https://www.vuemastery.com/courses/vue3-deep-dive-with-evan-you/vue3-overview",
   // "https://www.vuemastery.com/courses/vue3-typescript/why-vue-&-typescript",
   // "https://www.vuemastery.com/courses/typescript-friendly-vue3/introduction-to-the-script-setup-syntax",
