+++ index.js
@@ -2,4 +2,6 @@ require("dotenv").config();
 const app = require("./src/app");
 const courses = require("./courses");
 
-app.run(courses);
+const type = process.argv[2] || "all";
+
+app.run(courses, type);
