--- indexComplete.js
@@ -1,5 +1,0 @@
-require("dotenv").config();
-const appComplete = require("./src/appComplete");
-const courses = require("./courses");
-
-appComplete.run(courses);
